# documentacao/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Carango Bom API",
    "description": "Essa é a documentação da API do projeto Carango Bom.",
    "contact": {
      "email": "bizdev.tech@gmail.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
...
  "paths": {
...
    "/api/veiculos": {
      "post": {
        "tags": [
          "Veículo"
        ],
        "summary": "Cadastra um novo veículo e atribui um ID a ele",
        "operationId": "cadastraVeiculo",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VeiculoInputDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Novo veículo cadastrado",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/RespostaHttp"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "dados": {
                          "type": "object",
                          "$ref": "#/components/schemas/Veiculo"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "ID da marca para vincular veículo não existe",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MensagemHttp"
                }
              }
            }
          },
          "401": {
            "description": "Token JWT inválido ou ausente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MensagemHttp"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Veículo"
        ],
        "summary": "Lista todos os veículos",
        "operationId": "listaVeiculos",
        "responses": {
          "200": {
            "description": "Devolve um array de veículos",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "dados": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Veiculo"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/veiculos/{veiculoId}": {
      "put": {
        "tags": [
          "Veículo"
        ],
        "summary": "Altera um veículo existente",
        "operationId": "alteraVeiculo",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "veiculoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      }
                    }
                  },
                  {
                    "$ref": "#/components/schemas/VeiculoInputDto"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Veículo alterado",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/RespostaHttp"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "dados": {
                          "type": "object",
                          "$ref": "#/components/schemas/Veiculo"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Token JWT inválido ou ausente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MensagemHttp"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Veículo"
        ],
        "summary": "Excluir veículo existente",
        "operationId": "excluiVeiculo",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "veiculoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Veículo excluído"
          },
          "401": {
            "description": "Token JWT inválido ou ausente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MensagemHttp"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "RespostaHttp": {
        "type": "object",
        "properties": {
          "dados": {
            "type": "object"
          }
        }
      },
      "MensagemHttp": {
        "type": "object",
        "properties": {
          "mensagem": {
            "type": "string",
            "description": "Usado quando a API precisa informar mensagens de erro"
          }
        }
      },
...
      "Marca": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "UUID"
          },
          "nome": {
            "type": "string"
          }
        }
      },
      "Veiculo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "UUID"
          },
          "modelo": {
            "type": "string"
          },
          "ano": {
            "type": "integer"
          },
          "valor": {
            "type": "number"
          },
          "imagemUrl": {
            "type": "string",
            "format": "url",
            "description": "Endereço onde a imagem pode ser visualizada"
          },
          "marca": {
            "$ref": "#/components/schemas/Marca"
          }
        }
      },
...
      "VeiculoInputDto": {
        "type": "object",
        "properties": {
          "modelo": {
            "type": "string"
          },
          "ano": {
            "type": "integer"
          },
          "valor": {
            "type": "number"
          },
          "imagemUrl": {
            "type": "string",
            "format": "url",
            "description": "Endereço onde a imagem pode ser visualizada"
          },
          "marcaId": {
            "type": "string",
            "format": "UUID",
            "description": "ID da marca ao qual o veículo está vinculado"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}